+++ cratedb_sqlparse_js/cratedb_sqlparse/index.js
@@ -1,3 +1,3 @@
 import {sqlparse, Statement} from "./parser.js";
 export {sqlparse, Statement};
-export const __cratedb_version__ = "5.7.5"
+export const __cratedb_version__ = "5.8.3"